build gateway into core
