--- backend/gateway/Dockerfile.prod
@@ -1,15 +1,0 @@
-FROM node:16-alpine
-
-RUN apk add curl
-HEALTHCHECK CMD curl -f http://localhost:4000/graphql?query=%7B__typename%7D || exit 1
-
-WORKDIR /project/app
-
-ENV NODE_ENV=production
-
-COPY ./backend/gateway ./
-COPY ./backend/shared ../shared
-RUN cd ../shared && npm ci && npm run build && cd ../app && npm install ../shared && npm ci && npm run build
-
-
-CMD npm start
