--- backend/gateway/package.json
@@ -1,31 +1,0 @@
-{
-  "name": "dsek-gateway",
-  "version": "1.0.0",
-  "description": "The gateway for the GraphQL API.",
-  "main": "./dist/gateway.js",
-  "scripts": {
-    "build": "tsc",
-    "dev": "ts-node-dev --respawn --transpile-only --watch src src/gateway.ts",
-    "start": "node dist/gateway.js",
-    "test": "echo \"Error: no test specified\" && exit 1"
-  },
-  "author": "",
-  "license": "ISC",
-  "dependencies": {
-    "@apollo/gateway": "^0.28.1",
-    "@types/jsonwebtoken": "^8.5.1",
-    "@types/wait-on": "^5.3.0",
-    "apollo-server-core": "^2.18.2",
-    "apollo-server-express": "^2.18.2",
-    "axios": "^0.21.1",
-    "axios-retry": "^3.2.4",
-    "dsek-shared": "file:../shared",
-    "express": "^4.17.1",
-    "graphql": "^15.3.0",
-    "jsonwebtoken": "^8.5.1",
-    "typescript": "^4.0.3"
-  },
-  "devDependencies": {
-    "ts-node-dev": "^1.0.0"
-  }
-}
