+++ backend/services/core/src/gateway.ts
@@ -1,0 +1,86 @@
+import jwt from 'jsonwebtoken';
+import axios from 'axios';
+import axiosRetry from 'axios-retry';
+
+axiosRetry(axios, { retries: 10, retryDelay: (count) => count * 500 });
+
+/*
+ * The following interfaces are what data the keycloak token includes.
+ *
+ * OpenIdToken is defined at https://openid.net/specs/openid-connect-core-1_0.html#IDToken.
+ * KeycloakToken is created by inspecting a token from our keycloak instance.
+ *
+ * Most of these field most likely wont be used.
+ */
+interface OpenIdToken {
+  iss: string,
+  sub: string, // User id
+  aud: string[],
+  exp: number,
+  iat: number,
+  auth_time?: number,
+  nonce?: string,
+  acr?: string,
+  amr?: string[],
+  azp?: string,
+}
+
+interface KeycloakToken {
+  jti?: string,
+  nbf?: number,
+  typ?: string,
+  session_state?: string,
+  'allowed-origins'?: string[],
+  realm_access?: {
+    roles?: string[], // What roles a user has
+  },
+  resource_access?: {
+    'realm-management'?: {
+      roles?: string[],
+    },
+    account?: {
+      roles?: string[],
+    },
+  },
+  scope?: string,
+  email_verified?: boolean,
+  name?: string,
+  preferred_username?: string, // Most likely their student id
+  given_name?: string,
+  family_name?: string,
+  email?: string,
+  group: string[],
+}
+
+let pemCache: string | undefined;
+const pemCacheTtl = 60 * 1000;
+const keycloakAddress = 'https://portal.dsek.se/auth/realms/dsek/';
+
+type Token = KeycloakToken & OpenIdToken | undefined
+
+/**
+ * turns dsek.sexm.kok.mastare into ['dsek', 'dsek.sexm', 'dsek.sexm.kok', 'dsek.sexm.kok.mastare']
+ * @param id the key of the position
+ * @returns return a list of roles part of the position
+ */
+export function getRoleNames(id: string): string[] {
+  const parts = id.split('.');
+  return [...Array(parts.length).keys()].map((i) => parts.slice(0, i + 1).join('.'));
+}
+
+export const verifyAndDecodeToken = async (token: string): Promise<Token> => {
+  let pem = pemCache; // To avoid race conditions
+  if (!pem) {
+    const res = await axios.get(keycloakAddress);
+    const key = res.data.public_key;
+    pemCache = `-----BEGIN PUBLIC KEY-----\n${key}\n-----END PUBLIC KEY-----`;
+    pem = pemCache;
+    setTimeout(() => { pemCache = undefined; }, pemCacheTtl);
+  }
+
+  try {
+    return jwt.verify(token, pem) as KeycloakToken & OpenIdToken;
+  } catch (e) {
+    return undefined;
+  }
+};
