+++ docker-compose.prod.yml
@@ -11,10 +11,6 @@ services:
     restart: always
     build:
       dockerfile: ./backend/services/core/Dockerfile.prod
-  gateway:
-    restart: always
-    build:
-      dockerfile: ./backend/gateway/Dockerfile.prod
   frontend:
     restart: always
     build:
